validate proxy mesh config
