+++ model/BUILD
@@ -17,6 +17,8 @@ go_library(
         "@com_github_golang_glog//:go_default_library",
         "@com_github_golang_protobuf//jsonpb:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
+        "@com_github_golang_protobuf//ptypes:go_default_library",
+        "@com_github_golang_protobuf//ptypes/duration:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
         "@io_istio_api//:go_default_library",
     ],
@@ -41,6 +43,7 @@ go_test(
         "config_test.go",
         "mock_config_gen_test.go",
         "service_test.go",
+        "validation_test.go",
     ],
     library = ":go_default_library",
     deps = [
